getFormattedWidgetColumns function
